feat: hacky network and notification widget and caffeine widget
Network and notification widgets are just buttons to open nmtui and
swaync respectively... what they are doing in my waybar right now
anyways. Now I *should* not need waybar anymore.

Caffeine widget does what there is to do, set an inhibitor.

Restructured the project so that icons are in src/, screenshots are in
the renamed screenshots folder